% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{acc_rate}
\alias{acc_rate}
\title{Acceptance Rate Calculator}
\usage{
acc_rate(chain)
}
\arguments{
\item{chain}{vector of mcmc values}
}
\description{
Calculates the acceptance rate of an MCMC chain by looking at the number of repeats.
}
