% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{cbs}
\alias{cbs}
\title{Cubic B-spline with evenly spaced knots}
\usage{
cbs(x, n_int_knots)
}
\arguments{
\item{x}{a vector of evaluation points.}

\item{n_int_knots}{the number of internal knots.}
}
\description{
Evaluates the B-spline expansion at a point or vector of points. These restrictions can take 
advantage of polynomial expressions that are much faster to compute.
}
