% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countrydata}
\alias{countrydata}
\title{Our Country Metadata
Our personal country-metdata}
\format{
A data frame with 261 rows and 74 variables:
\describe{
  \item{Region Name}{Coming soon}
  ...
}
}
\usage{
countrydata
}
\description{
Our Country Metadata
Our personal country-metdata
}
\keyword{datasets}
