% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpub_savejson.R
\name{cpub_savejson}
\alias{cpub_savejson}
\title{Highcharts options export function}
\usage{
cpub_savejson(plot1, filename)
}
\arguments{
\item{plot1}{HIghcharter HMTL widget}

\item{filename}{Name of the file to save}
}
\description{
Highcharts options export function
}
\examples{
library(highcharter)
plot<- hchart(mtcars$mpg)
cpub_savejson(plot, "mtcars")


}
