% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpub_sentiment.R
\name{cpub_sentiment}
\alias{cpub_sentiment}
\title{Augment a data frame to include text sentiments from bing, afinn, NRC}
\usage{
cpub_sentiment(df, text_column)
}
\arguments{
\item{df}{Data frame to work with}

\item{text_column}{Column of text}
}
\description{
Augment a data frame to include text sentiments from bing, afinn, NRC
}
\examples{
library(dplyr)
text <- c("Because I could not stop for Death -","He kindly stopped for me -","The Carriage held but just Ourselves -","and Immortality")
text_df <- tibble(line = 1:4, text = text)
cpub_sentiment(text_df, text)

}
