% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{det_sympd}
\alias{det_sympd}
\title{Determinant of Symmetric Positive Definite Matrix}
\usage{
det_sympd(x, log = FALSE)
}
\arguments{
\item{x}{a numeric matrix.}

\item{log}{logical; TRUE returns the log determinant.}
}
\description{
Computes the (log) determinant of a SymPD matrix using the Cholesky factorization.
}
