% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_functions.R
\name{dinvgamma}
\alias{dinvgamma}
\title{Inverse Gamma Density Function}
\usage{
dinvgamma(x, shape, rate, scale = 1/rate, log = FALSE)
}
\arguments{
\item{x}{vector of nonnegative values.}

\item{shape, scale}{shape and scale parameters. Must be stricly positive.}

\item{rate}{alternative way to specify scale.}

\item{log}{logical; if TRUE, density calculations are computed on the log scale.}
}
\description{
Computes the (log) density of the inverse gamma distribution using
either the scale or rate parametrization. The scale parametrization has
expected value of scale/(shape - 1)
}
