% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_functions.R
\name{dmnorm}
\alias{dmnorm}
\title{Multivariate Normal Density Function}
\usage{
dmnorm(x, mu, cov, prec, unnorm = FALSE, log = FALSE)
}
\arguments{
\item{x}{vector or matrix of values. Multiple observations should correspond to rows.}

\item{mu}{mean vector.}

\item{cov, prec}{covariance or precision matrix. Only specify one.}

\item{unnorm}{logical; TRUE ommits all terms not affected by x.}

\item{log}{logical; if TRUE, density calculations are computed on the log scale.}
}
\description{
Computes the (log) density of the multivariate normal distribution
using either the covariance or precision parametrization.
}
