% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{globalMapData}
\alias{globalMapData}
\title{Global Map List Object
Downloaded highcharts map-data objects}
\format{A list
\describe{
  \item{Global Map Data}{Coming soon}
  ...
}}
\usage{
globalMapData

globalMapData
}
\description{
Global Map List Object
Downloaded highcharts map-data objects

European Map List Object
Downloaded highcharts map-data objects
}
\keyword{datasets}
