% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctzn.R
\name{hc_theme_ctzn}
\alias{hc_theme_ctzn}
\title{CPUB theme for highcharts}
\usage{
hc_theme_ctzn(...)
}
\arguments{
\item{...}{Named argument to modify the theme}
}
\description{
Out theme, Based on Highcharter Google theme 1.
}
\examples{

highcharts_demo() \%>\%
  hc_add_theme(hc_theme_ctzn())

}
