% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctzn2.R
\name{hc_theme_ctzn2}
\alias{hc_theme_ctzn2}
\title{CPUB theme for highcharts}
\usage{
hc_theme_ctzn2(...)
}
\arguments{
\item{...}{Named argument to modify the theme}
}
\description{
Our theme, Based on Highcharter Tufte theme.
}
\examples{

highcharts_demo() \%>\%
  hc_add_theme(hc_theme_ctzn2())

}
