% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{is_monotone}
\alias{is_monotone}
\title{Check is vector is monotone}
\usage{
is_monotone(x, strict)
}
\arguments{
\item{x}{numeric vector}

\item{strict}{logical; if true then strict monotonicity is determined}
}
\description{
Determines if the sequence of vector elements are increasing or decreasing with the option to
specify strictness.
}
