% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{Mahalanobis}
\alias{Mahalanobis}
\title{Fast Mahalanobis Distance Computation}
\usage{
Mahalanobis(x, center, cov, prec)
}
\arguments{
\item{x}{a vector or matrix. Multiple observations should correspond to rows.}

\item{center}{the mean or centering vector}

\item{cov, prec}{covariance or precision matrices. Only provide one.}
}
\description{
Quickly evaluates Mahalanobis distance given either a covariance or precision matrix.
}
