% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{monotonize}
\alias{monotonize}
\title{Project a vector into monotone space}
\usage{
monotonize(x, type = "increasing", forced = NULL)
}
\arguments{
\item{x}{non-empty numeric vector}

\item{type}{character indicating if the function is increasing or decreasing}

\item{forced}{numeric vector of indeces specifying points that cannot be moved}
}
\description{
Makes a vector monotone. (REFERENCE) proposed a method to project a vector onto monotone space.
}
