% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_highchart.R
\name{plot_highchart}
\alias{plot_highchart}
\title{Highcharts for SJPlot outputs from a regression}
\usage{
plot_highchart(model, terms, type, colors, size)
}
\arguments{
\item{model}{An model object}

\item{terms}{Desired terms}

\item{type}{Optional, defaults to 'eff' but can also be 'pred' if specified.}

\item{colors}{optional vectors of colors c("#9A32CD", "#FF3030", "#7FFFD4")}
}
\description{
Highcharts for SJPlot outputs from a regression
}
\examples{
data(efc)
efc$c161sex <- to_factor(efc$c161sex)
fit <- lm(neg_c_7 ~ c12hour + barthtot * c161sex, data = efc)

plot_model(fit, type = "pred", terms = c("barthtot [30,50,70]", "c161sex"))
plot_highchart(fit,  terms = c("c161sex", "barthtot [30,50,70]"), color = c("red", "pink", "blue"))

}
