% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{stack}
\alias{stack}
\title{Stack a List of Vectors or Matrices}
\usage{
stack(x)
}
\arguments{
\item{x}{list of vectors or matrices with the same number of columns}
}
\value{
a list containing the stacked object and a vector of indices the specify where
  each of the original elements begins.
}
\description{
Takes a list of vectors or matrices and returns one vector or matrix along with
a vector that specifies the first indices corresponding to the original elements.
}
