% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{stack_Matrix}
\alias{stack_Matrix}
\title{Stack A List of Matrices}
\usage{
stack_Matrix(x)
}
\arguments{
\item{x}{list of matrices.}
}
\description{
Stack a list of matrices together into one tall matrix. This is a simplified version of the
stack function.
}
