% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabs.R, R/tabs2.R
\name{tabs}
\alias{tabs}
\title{Mouseover Tabsets Optimized for Distill Rmd Articles}
\usage{
tabs(tabID, list_of_tabs, list_of_widgets, height)

tabs(tabID, list_of_tabs, list_of_widgets, height)
}
\arguments{
\item{tabID}{A number, 1-9 (needs a unique ID to distinguish from other tabsets
)}

\item{list_of_tabs}{A list of named vectors with a tab title & url for content}

\item{height}{Desired height.}

\item{icon}{optional; takes a fontawesome identifier to add an icon}
}
\description{
Our tabets, made for Distill articles in R markdown. Returns the HTML code needed to load various scripts used and takes a list of tab names/iframe content as arguments.

Our tabets, made for Distill articles in R markdown. Returns the HTML code needed to load various scripts used and takes a list of tab names/iframe content as arguments.
}
\examples{

tabs(
tabID = 2,
list_of_tabs = list(
  c(title = "Income Inequality", icon = "fas fa-bezier-curve", url = "https://www.onmundo.org/cpub/visualization/viz/global_adolescent_birth_rate_GINIBubbleD.html"),
  c(title = "GDP", url = "https://www.onmundo.org/cpub/visualization/viz/global_adolescent_birth_rate_GDPBubbleD.html"),
  c(title = "Gender Inequality", url = "https://www.onmundo.org/cpub/visualization/viz/global_adolescent_birth_rate_GIIBubbleD.html"),
  c(title = "Human Development", url = "https://www.onmundo.org/cpub/visualization/viz/global_adolescent_birth_rate_HDIBubbleD.html")
),
height = 500)



tabs(
tabID = 2,
list_of_tabs = list(
  c(title = "Income Inequality", icon = "fas fa-bezier-curve", url = "https://www.onmundo.org/cpub/visualization/viz/global_adolescent_birth_rate_GINIBubbleD.html"),
  c(title = "GDP", url = "https://www.onmundo.org/cpub/visualization/viz/global_adolescent_birth_rate_GDPBubbleD.html"),
  c(title = "Gender Inequality", url = "https://www.onmundo.org/cpub/visualization/viz/global_adolescent_birth_rate_GIIBubbleD.html"),
  c(title = "Human Development", url = "https://www.onmundo.org/cpub/visualization/viz/global_adolescent_birth_rate_HDIBubbleD.html")
),
height = 500)


}
