% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toc.R
\name{toc_function}
\alias{toc_function}
\title{Table of contents for Distill Rmd Articles}
\usage{
toc_function(toc)
}
\arguments{
\item{list_of_tabs}{A list of named vectors with a tab title & url for content}
}
\description{
Our  table of contents, made for Distill articles in R markdown. Returns the HTML code needed to load various scripts used and takes a list of tab names/iframe content as arguments.
}
\examples{

toc <- list(
 c(title = "1. Overview", icon = "fas fa-bezier-curve", description = "this is the overview of our document"),
 c(title = "2. This 2nd section", icon = "fas fa-bezier-curve", description = "second section covers other stuff"),
c(title = "2a. Overview again", icon = "fas fa-bezier-curve", description = "this is the overview") )

}
