% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{unstack}
\alias{unstack}
\title{Unstack a Vector or Matrix}
\usage{
unstack(x, inds, n)
}
\arguments{
\item{x}{vector or matrix.}

\item{inds}{vector indicating how to split x. If not provided then it will
automatically attempt to split into equal parts.}

\item{n}{indicates the number of elements to split into.
Only needed when inds is not provided.}
}
\value{
list containing the split elements.
}
\description{
Takes a vector or matrix and an index vector and returns a list containing the pieces.
}
