% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjugate.R
\name{update_normal_invgamma}
\alias{update_normal_invgamma}
\title{Conjugate Normal - Inverse Gamma Update}
\usage{
update_normal_invgamma(y, a, b, mu, R, R_inv)
}
\arguments{
\item{y}{vector of values at the likelihood level.}

\item{a}{prior shape value for inverse gamma.}

\item{b}{prior SCALE value for inverse gamma.}

\item{mu}{mean vector for multivariate normal.}

\item{R}{correlation matrix for multivariate normal.}

\item{R_inv}{scaled precision matrix, alternative specification to R.}
}
\description{
Generates a value from the posterior distribution in the case where there
is a multivariate normal likelihood and an inverse gamma prior.\cr \cr
Argument model:\cr
y ~ Nn(mu, sig2*R)\cr
sig2 ~ IG(a,b)\cr
}
