% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjugate.R
\name{update_normal_normal}
\alias{update_normal_normal}
\title{Conjugate Multivariate Normal - Multivariate Normal Update}
\usage{
update_normal_normal(y, X, mu, Sig, V, Sig_inv, V_inv)
}
\arguments{
\item{y}{vector of values at the likelihood level.}

\item{X}{fixed design matrix in likelihood.}

\item{mu}{prior mean vector.}

\item{Sig, Sig_inv}{likelihood covariance/precision matrix.}

\item{V, V_inv}{prior covariance/precision matrix.}
}
\description{
Generates a value from the posterior distribution in the case where
there is a multivariate normal likelihood and a multivariate normal prior.\cr \cr
Argument Model:\cr
y ~ Nn(X*beta, Sig)\cr
beta ~ Np(mu, V)\cr
}
