% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_summary.R
\name{wbPlots}
\alias{wbPlots}
\title{Making plots from a world bank indicator}
\usage{
wbPlots(indicator)
}
\arguments{
\item{indicator}{This should be an indicator, eg. "SH.ALC.PCAP.LI"}
}
\description{
A function that takes a world band indicator and subsets to cross sectional data from the latest observation for each country. Returns 5 plots: Top Countries, Bottom Countries, Continuous Map, Discrete Map & A topogram, in that order.
}
