% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztabs.R
\name{ztabs}
\alias{ztabs}
\title{Tabsets for RMD}
\usage{
ztabs(tab_titles, tab_contents, name)
}
\arguments{
\item{tab_titles}{The titles of the tabs (must be a list and same length/order as content)}

\item{tab_contents}{The contents of the tabs (must be a list and same length/order as titles)}

\item{name}{Unique name for the tabset}
}
\description{
Tabsets for RMD
}
\examples{

ztabs(name = 'name',  tab_contents= list('panel1 content', 'panel2 content')  ,  tab_titles =list('title1','title2') )

You can also add in little icons with fontawesome

 tab_titles= list(tags$div(fa(name = 'venus-mars', fill="steelblue"), "Sex")

** note...make sure you have access to the proper styles file.
}
