\name{ztools-package}
\alias{ztools-package}
\alias{ztools}
\docType{package}
\title{
\packageTitle{ztools}
}
\description{
\packageDescription{ztools}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ztools}
\packageIndices{ztools}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ztools}

Maintainer: \packageMaintainer{ztools}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
