% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpub_sentiment.R
\name{cpub_sentiment}
\alias{cpub_sentiment}
\title{Augment a data frame to include text sentiments from  afinn, NRC}
\usage{
cpub_sentiment(df, text_column)
}
\arguments{
\item{df}{Data frame to work with}

\item{text_column}{Column of text}
}
\description{
GI, QDAP
Harvard-IV dictionary (GI) http://www.wjh.harvard.edu/~inquirer/spreadsheet_guide.htm
QDAP dictionary from the package qdapDictionaries
}
\examples{
library(dplyr)
text <- c("I hate everything smells foul and ugly and mean and evil","Kind love peace harmony","The Carriage held but just Ourselves -","and Immortality")
text_df <- tibble(line = 1:4, text = text)
cpub_sentiment(text_df, text)

}
